api/pyproject.toml
[tool.poetry]
name = "apps.gennet.ai"
description = "Gennet AI Backend API with FastAPI"
version = "0.2.0"
authors = ["Abubakar Abubakar Yusif <abubakaryusuf951@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.7.0"
fastapi = "^0.110.2"
gradio_client = "^0.15.0"
httpx = "^0.27.0"
pydantic = {extras = ["dotenv", "email"], version = "^2.7.1"}
pydantic_settings = "^2.2.1"
pytest = "^8.1.2"
PyYAML = "^6.0.1"
scidownl = "^1.0.2"
starlette = "^0.37.2"
gunicorn = "^21.2.0"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
requests = "^2.32.1"
pkce = "^1.0.3"
alembic = "^1.13.1"
annotated-types = "^0.7.0"
anyio = "^4.3.0"
asyncpg = "^0.29.0"
bcrypt = "^4.1.3"
cffi = "^1.16.0"
cryptography = "^42.0.7"
dnspython = "^2.6.1"
email-validator = "^2.1.1"
greenlet = "^3.0.3"
idna = "^3.7"
mako = "^1.3.5"
markupsafe = "^2.1.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pycparser = "^2.22"
pydantic-core = "^2.18.2"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
sniffio = "^1.3.1"
sqlalchemy = "^2.0.30"
typing-extensions = "^4.11.0"
xmltodict = "^0.13.0"
jinja2 = "^3.1.4"
pytest-asyncio = "^0.23.7"
sentry-sdk = "^2.2.0"
scalar-fastapi = "^1.0.2"
tenacity = "^8.2.3"
emails = "^0.6"
psycopg = {extras = ["binary"], version = "^3.1.13"}
sqlmodel = "^0.0.16"
# Pin bcrypt until passlib supports the latest
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["auth", "db", "h2ogpt", "medpub", "."]
asyncio_mode = "auto"

[tool.poetry.extras]
